+++ src/lib/types_test.rs
@@ -433,7 +433,7 @@ fn task_extend_both_have_misc_data() {
     base.command = Some("test1".to_string());
     base.disabled = Some(false);
     base.private = Some(false);
-    base.watch = Some(false);
+    base.watch = Some(TaskWatchOptions::Boolean(false));
     base.script = Some(vec!["1".to_string(), "2".to_string()]);
 
     let extended = Task {
@@ -445,7 +445,7 @@ fn task_extend_both_have_misc_data() {
         workspace: None,
         disabled: Some(true),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
         force: Some(true),
@@ -509,7 +509,7 @@ fn task_extend_both_have_misc_data() {
     assert_eq!(base.command.unwrap(), "test1");
     assert!(base.disabled.unwrap());
     assert!(base.private.unwrap());
-    assert!(base.watch.unwrap());
+    assert_eq!(base.watch.unwrap(), TaskWatchOptions::Boolean(true));
     assert!(base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 0);
     assert_eq!(base.alias.unwrap(), "alias2");
@@ -527,7 +527,7 @@ fn task_extend_extended_have_all_fields() {
         workspace: None,
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
         force: Some(true),
@@ -563,7 +563,7 @@ fn task_extend_extended_have_all_fields() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
-        watch: Some(false),
+        watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -596,7 +596,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -626,7 +626,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -656,7 +656,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -725,7 +725,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.workspace.unwrap());
     assert!(base.disabled.unwrap());
     assert!(!base.private.unwrap());
-    assert!(!base.watch.unwrap());
+    assert_eq!(base.watch.unwrap(), TaskWatchOptions::Boolean(false));
     assert_eq!(base.condition_script.unwrap().len(), 1);
     assert!(!base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 1);
@@ -768,7 +768,7 @@ fn task_extend_clear_with_no_data() {
         workspace: Some(false),
         disabled: Some(true),
         private: Some(false),
-        watch: Some(false),
+        watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -801,7 +801,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -831,7 +831,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -861,7 +861,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -938,7 +938,7 @@ fn task_extend_clear_with_all_data() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
-        watch: Some(false),
+        watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -971,7 +971,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1001,7 +1001,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1031,7 +1031,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1138,7 +1138,7 @@ fn task_get_normalized_task_undefined() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
         force: None,
@@ -1203,7 +1203,10 @@ fn task_get_normalized_task_undefined() {
     assert!(!normalized_task.workspace.unwrap());
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
-    assert!(normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(true)
+    );
     assert!(!normalized_task.force.unwrap_or(false));
     assert_eq!(normalized_task.alias.unwrap(), "alias");
     assert_eq!(normalized_task.linux_alias.unwrap(), "linux");
@@ -1243,7 +1246,7 @@ fn task_get_normalized_task_with_override_no_clear() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1272,7 +1275,7 @@ fn task_get_normalized_task_with_override_no_clear() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1348,7 +1351,10 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
-    assert!(!normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(false)
+    );
     assert_eq!(normalized_task.condition_script.unwrap().len(), 1);
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
@@ -1391,7 +1397,7 @@ fn task_get_normalized_task_with_override_clear_false() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1419,7 +1425,7 @@ fn task_get_normalized_task_with_override_clear_false() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string()]),
@@ -1499,7 +1505,10 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
-    assert!(!normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(false)
+    );
     assert_eq!(normalized_task.condition_script.unwrap().len(), 2);
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
@@ -1537,7 +1546,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1628,7 +1637,10 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert_eq!(normalized_task.command.unwrap(), "command");
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
-    assert!(normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(true)
+    );
     assert!(!normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 0);
     assert_eq!(normalized_task.cwd.unwrap(), "cwd".to_string());
@@ -1661,7 +1673,7 @@ fn task_get_normalized_task_with_override_clear_true() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),